+++ packages/carp_survey_package/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 1.1.0
+## 1.3.0
 
-* update to `carp_mobile_sensing` v. 1.1.0
+* update to `carp_mobile_sensing` v. 1.3.0
 * update to the new `UserTask` model where the app (and not this package) is responsible for showing the user interface (`Widget`) of the survey task.
+* upgrade of permission_handler plugins
 
 ## 0.40.1